Mixture model classes
This is an experimental file that has not been finished! It will compile, but its member functions and instance variables are being written and tested, so they either do not exist yet in this file, or are liable to do anything. This file is being uploaded to keep track of my progress.